+++ public/index.html
@@ -148,6 +148,23 @@
                         </div>
                     </div>
                 </div>
+                <!-- <div class="modal fade" id="searchResultsModal" tabindex="-1" role="dialog"
+                aria-labelledby="searchResultsModalLabel" aria-hidden="true">
+                <div class="modal-dialog" role="document">
+                    <div class="modal-content">
+                        <div class="modal-header">
+                            <h5 class="modal-title" id="searchResultsModalLabel">New BFFS:</h5>
+                        </div>
+                        <div class="modal-body" id="results-modal-body">
+                            
+                        </div>
+                        <div class="modal-footer">
+                            <button type="button" class="btn btn-secondary"
+                                data-dismiss="modal">Close</button>
+                        </div>
+                    </div>
+                </div>
+                 -->
                 <div class="row">
                     <div class="col-md-12" id="search-container">
                     </div>
@@ -171,6 +188,7 @@
         <script src="https://unpkg.com/@petfinder/petfinder-js/dist/petfinder.min.js"></script>
         <script src="/assets/js/index.js" type="text/javascript"></script>
         <script src="/assets/js/newuser.js" type="text/javascript"></script>
+        <script src="/assets/js/search.js" type="text/javascript"></script>
 
 
     </body>
